pyproject.toml
[project]
name = "agent-scraper"
version = "0.1.0"
description = "AI-driven web scraper using browser automation and LLM"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "browser-use>=0.12",
    "fastapi>=0.115",
    "openai>=1.50",
    "playwright>=1.48",
    "pydantic>=2.10",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_scraper", "src/autoscraper", "src/server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
